Copytruncate for postgresql log rotation
The postgres logging collector keeps logging to the same inode, so if we
let logrotate just rename the file, postgres keeps sending logs to the
old logfile and postgresql.log stays empty.

Here, we tell logrotate to copy and truncate instead of renaming the
file.

The alternative is to use PostgreSQL's built-in log rotation features.
However, that would mean using a rotating pattern like `postgresql-$day.log`.
In turn, fluentd would have to follow this rotation - which also isn't great, if possible at all.