Allow tags to be configured
Patroni tags
(https://patroni.readthedocs.io/en/latest/SETTINGS.html#tags) allow
certain nodes to be treated differently to the others. These can be
controlled using node-level chef attributes, for example to prevent a
node becoming the master.

Due to how we generate the config file this would likely have already
worked, but this change makes it explicit and documents it with tests.