Revert changes to template patroni.systemd.erb
This template does not have access to the recipe's local variable
"patroni_conf", and the only values it needs as input from the
Chef::Node object are not affected by the deep_merging of secrets
with node attributes.  So the "node" hash gives the same answer as
"patroni_conf" would have.