Run VACUUM instead of ANALYZE after Patroni failover
This patch implements the 2nd corrective action described here:

https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2937#note_445840087

Historically, the "db-analyze.sh" script was added as a task for Patroni to run
as a follow-up to failover events.  The goal was to prevent false alarms
from our "TooManyDeadTuples" alert, which is derived from the live versus dead
tuple estimates presented in the "pg_stat_user_tables" view.  That prior work
is documented here:

https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/5841

Unfortunately, the choice to run ANALYZE with a smaller-than-normal sample size has
the unintended side-effect of significantly lowering the quality of the optimizer statistics,
which can sometimes severely impact query performance.  This is what caused the recent
incident (issue 2937, linked above), for which this patch is a corrective action.

So we need an alternative solution that avoids false alarms for the "TooManyDeadTuples"
alert without risk of introducing a real performance regression.  Fortunately,
the live/dead tuple count estimates are also calculated by VACUUM in a manner
similar to how ANALYZE does.  This means we can run VACUUM instead of ANALYZE to
calculate fresh live/dead tuple statistics without also replacing the column-level
optimizer statistics.

For additional background on the problem with using ANALYZE, see:
* Terse problem summary:
  https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2937#note_445229140
* Detailed walk-through of the underlying mechanism:
  https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2937#note_445326510