Avoid writing secrets as persistent node attributes
Running deep_merge! was writing secrets to the node attributes,
which exposes them and leaks them across runs.

Rename variable "node" to "patroni_conf", so foodcritic stops
treating it as a Chef::Node object instead of as a plain hash.

Revert changes to template patroni.systemd.erb.

This template does not have access to the recipe's local variable
"patroni_conf", and the only values it needs as input from the
Chef::Node object are not affected by the deep_merging of secrets
with node attributes.  So the "node" hash gives the same answer as
"patroni_conf" would have.

Disable unnecessary code style check.