ci: use specific version of ubuntu 16.04
Background
---
Ubuntu 16.04 is deprecated, however we are still [depending on
it](gitlab-com/gl-infra&231) so we
need to keep testing the cookbook with this version to make sure they
work. GCP deprecated all of the Ubuntu 16.04 and kitchen only looks up
for non-deprecated images.

Solution
---
Use a specific version of ubuntu that is deprecatd.

reference gitlab-com/gl-infra/infrastructure#14606